requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
eventlet==0.33.3
Werkzeug==2.3.7
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
adafruit-circuitpython-dht==3.7.3
tensorflow==2.12.0
python-dateutil==2.8.2
schedule==1.2.0